Move Basic inline functions to basic-inl.h.
Also add newlines to the end of each file.
